+++ src/main/java/org/matrix/izumbankapp/controller/AccountController.java
@@ -3,7 +3,6 @@ package org.matrix.izumbankapp.controller;
 import org.matrix.izumbankapp.model.accounts.*;
 import org.matrix.izumbankapp.model.auth.AccountStatusUpdate;
 import org.matrix.izumbankapp.model.auth.ResponseDto;
-import org.matrix.izumbankapp.model.deposits.DepositRequest;
 import org.matrix.izumbankapp.scheduler.DepositScheduler;
 import org.matrix.izumbankapp.service.AccountService;
 import jakarta.validation.Valid;
@@ -41,11 +40,6 @@ public class AccountController {
         return ResponseEntity.status(HttpStatus.OK).body(accountService.getAccountById(accountId));
     }
 
-    @GetMapping("/accountNumber")
-    public ResponseEntity<AccountResponse> getAccountByAccountNumber(@RequestParam String accountNumber) {
-        return ResponseEntity.status(HttpStatus.OK).body(accountService.getAccountByAccountNumber(accountNumber));
-    }
-
     @PostMapping
     public ResponseEntity<ResponseDto> createAccount(@Valid @RequestBody AccountCreateDto account) {
         return ResponseEntity.status(HttpStatus.CREATED).body(accountService.createAccount(account));
@@ -76,20 +70,6 @@ public class AccountController {
         return ResponseEntity.status(HttpStatus.OK).body(accountService.getBalance(accountNumber));
     }
 
-    @PostMapping("/transfer")
-    public ResponseEntity<ResponseDto> transferToAccount(@Valid @RequestBody TransferMoneyRequest transferMoneyRequest) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(accountService.transferToAccount(transferMoneyRequest));
-    }
-
-    @PostMapping("/withdrawal")
-    public ResponseEntity<ResponseDto> withdrawal(@Valid @RequestBody WithdrawalRequest withdrawalRequest) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(accountService.withdrawal(withdrawalRequest));
-    }
-
-    @PostMapping("/deposit")
-    public ResponseEntity<ResponseDto> deposit(@Valid @RequestBody DepositRequest depositRequest) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(accountService.createDepositAccount(depositRequest));
-    }
     @PostMapping("/testing")
     public ResponseEntity<ResponseDto> deposit() {
         return ResponseEntity.status(HttpStatus.CREATED).body(depositScheduler.calculateDepositInterest());
