+++ src/main/java/org/matrix/izumbankapp/controller/UserController.java
@@ -31,67 +31,32 @@ public class UserController {
     public Page<UserProfileDto> getUsersByFilter(UserProfileFilterDto filter, Pageable pageable) {
         return userService.findUsersByFilter(filter, pageable);
     }
-    /**
-     * Retrieves a user by their ID.
-     *
-     * @param userId the ID of the user to retrieve
-     * @return the user details as a ResponseEntity
-     */
+
     @GetMapping("/userId")
     public ResponseEntity<UserResponse> getUserById(@RequestParam Long userId) {
         return ResponseEntity.status(HttpStatus.OK).body(userService.getUserById(userId));
     }
-    /**
-     * Retrieves a user by their email.
-     *
-     * @param email the email of the user to retrieve
-     * @return the user details as a ResponseEntity
-     */
 
     @GetMapping("/email")
     public ResponseEntity<UserResponse> getUserByEmail(@Valid @RequestParam String email) {
         return ResponseEntity.status(HttpStatus.OK).body(userService.getUserByEmail(email));
     }
 
-    /**
-     * Retrieves all users.
-     *
-     * @return The list of user DTOs.
-     */
     @GetMapping
     public ResponseEntity<List<UserResponse>> getAllUser() {
         return ResponseEntity.status(HttpStatus.OK).body(userService.getAllUser());
     }
 
-    /**
-     * Creates a new user.
-     *
-     * @param user the user data transfer object.
-     * @return the response entity containing the response.
-     */
     @PostMapping
     public ResponseEntity<ResponseDto> createUser(@Valid @RequestBody UserCreateDto user) {
         return ResponseEntity.status(HttpStatus.CREATED).body(userService.addUser(user));
     }
 
-    /**
-     * Updates user's information.
-     *
-     * @param id The ID of the user to update.
-     * @param user The user to update.
-     * @return The response entity containing the response.
-     */
     @PutMapping
     public ResponseEntity<ResponseDto> updateUser(@RequestParam Long id, @Valid @RequestBody UserUpdateDto user) {
         return ResponseEntity.status(HttpStatus.OK).body(userService.updateUser(id, user));
     }
 
-    /**
-     * Deletes the user with the ID.
-     *
-     * @param id The ID of the user to delete.
-     * @return the response entity containing the response.
-     */
     @DeleteMapping
     public ResponseEntity<ResponseDto> deleteUserById(@RequestParam Long id) {
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body(userService.deleteUserById(id));
@@ -100,11 +65,6 @@ public class UserController {
     @GetMapping("/accounts/{accountNumber}")
     public ResponseEntity<UserAccountsResponse> readUserByAccountNumber(@PathVariable String accountNumber) {
         return ResponseEntity.status(HttpStatus.OK).body(accountService.getUserByAccountNumber(accountNumber));
-    }
-
-    @GetMapping("{userId}/accounts")
-    public ResponseEntity<List<AccountsUserResponse>> getAllAccountsByUserId(@PathVariable Long userId) {
-        return ResponseEntity.status(HttpStatus.OK).body(accountService.getAllAccountsByUserId(userId));
     }
 
 }
